Queue python
I didn't do it correctly based off of the instructions using len, or adding a range and I was struggling. However I did realize that I was able to repeat a certain part of the code in order to pop out eah name one by one to empty out the queue.